refactor: update navigation routes in onboarding_view.dart
This commit updates the navigation routes in the `onboarding_view.dart` file. It replaces the usage of `LoginPage.routeName` and `SignupPage.routeName` with `AppRoutes.login` and `AppRoutes.signUp` respectively. This change ensures consistency with the defined routes in the `app_routes.dart` file and improves the maintainability of the codebase.