refactor: update HeaderText component to accept nullable color
The code changes in header_text.dart modify the HeaderText component to accept a nullable color parameter. This allows for more flexibility in customizing the text color, as it can now be set to null to use the default color defined in the app's theme. This change improves the reusability and configurability of the HeaderText component.