refactor: Add date attribute to Comment and Post models
This refactor adds a new `date` attribute to the `Comment` and `Post` models. The `date` attribute represents the date and time when the comment or post was created. This addition enhances the functionality of the models by providing a timestamp for each comment and post, allowing for better sorting and organization of content based on creation time.