chore: update dependencies in pubspec.yaml
This code change updates the dependencies in the pubspec.yaml file. It includes updating the provider package to version 6.1.2 and adding the nested package with version 1.0.0. These updates ensure that the project is using the latest stable versions of the dependencies, improving compatibility and taking advantage of new features and bug fixes.