feat: add image and video picking functionality to AddPostPage
This code change adds the ability to pick images and videos from the device's gallery in the AddPostPage. It introduces the `_pickImage()` and `_pickVideo()` methods, which use the `ImagePicker` package to select an image or video file. The selected file is then stored in the `_image` or `_video` variables respectively. This enhancement allows users to include multimedia content in their posts, improving the overall user experience.